Removing networking
sal issues and that networking was totally screwed..  that and I'm not going to maintain such a 'thing'..  I'm not a network programmer and won't ever be.

Oh and fixed a few 64bit'isms

+++ src/kex/p_mobj.h
@@ -201,10 +201,10 @@ typedef struct mobj_s {
     mobjtype_t          type;
     mobjinfo_t*         info;    // &mobjinfo[mobj->type]
 
-    int                 tics;    // state tic counter
+    intptr_t            tics;    // state tic counter
     state_t*            state;
     dword               flags;
-    int                 health;
+    intptr_t            health;
 
     // [d64] alpha value for rendering
     int                 alpha;
